+++ src/sessions/base-session-handler.ts
@@ -115,7 +115,7 @@ export default abstract class BaseSessionHandler {
   }
 
   async endSession (session: IExtendedMediaSession, reason?: Constants.JingleReasonCondition): Promise<void> {
-    this.log('info', 'ending session', { conversationId: session.conversationId, reason });
+    this.log('info', 'ending session', { sessionId: session.id, conversationId: session.conversationId, reason });
 
     return new Promise<void>((resolve) => {
       session.once('terminated', (reason) => {